+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Welcome to the Github! If you are looking for the most recent stable version, th
 - End themed Ruined Portals now spawn in the End dimension!
   Special thanks to miguelforge for creating them!
 
+- Added Birch Outposts to any biome with "birch" in the name!
+
 - Added Oak Village that spawns in any forest category biome that isn't birch or dark forest!
 
 ##### Backend: